Revert "Standardize competitionId and shooterId to Long throughout"
This reverts commit 0ae0a15b74b33947fdde9edf1313cf5127d3153f.

+++ app/src/main/java/se/kjellstrand/webshooter/data/results/local/ResultsDao.kt
@@ -8,11 +8,11 @@ import androidx.room.Query
 @Dao
 interface ResultsDao {
     @Query("SELECT * FROM results WHERE competitionsId = :competitionId")
-    suspend fun getByCompetition(competitionId: Long): List<ResultEntity>
+    suspend fun getByCompetition(competitionId: Int): List<ResultEntity>
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertAll(results: List<ResultEntity>)
 
     @Query("DELETE FROM results WHERE competitionsId = :competitionId")
-    suspend fun deleteByCompetition(competitionId: Long)
+    suspend fun deleteByCompetition(competitionId: Int)
 }
